Split into hpp and cpp files
